add feature location

+++ ewm/src/main/java/com/github/artemlv/ewm/location/validation/ConstraintNotZeroValidation.java
@@ -5,9 +5,9 @@ import jakarta.validation.ConstraintValidatorContext;
 
 import java.util.Objects;
 
-public class ConstraintNotZeroValidation implements ConstraintValidator<ConstraintNotZero, Long> {
+public class ConstraintNotZeroValidation implements ConstraintValidator<ConstraintNotZero, Double> {
     @Override
-    public boolean isValid(Long number, ConstraintValidatorContext constraintValidatorContext) {
+    public boolean isValid(Double number, ConstraintValidatorContext constraintValidatorContext) {
         return Objects.nonNull(number) && number != 0;
     }
 }
